public/chatbot/requirement.txt
fastapi
uvicorn
pydantic
google-generativeai
python-dotenv
supabase
httpx